couple bug issues

+++ frontend/src/components/AdminPageComp/AdminAdmin.jsx
@@ -2,6 +2,8 @@ import { useState } from "react"
 import ViewPopup from "../ViewPopup";
 import axios from "axios";
 import SimpleAlert from "../SimpleAlert";
+import { confirmAlert } from 'react-confirm-alert';
+import 'react-confirm-alert/src/react-confirm-alert.css';
 
 const AdminAdmin = ({admin}) => {
 
@@ -20,14 +22,33 @@ const AdminAdmin = ({admin}) => {
     setTimeout(() => setShowAlert(false), 3000);
 };
 
-  const handleDelete = ()=>{
-    deleteAdmin(admin.userId);
-    handleAlert();
-    setTimeout(() => {
-      window.location.reload();
-    }, 3000);
-   
-  }
+
+  const handleDelete = () => {
+    confirmAlert({
+      title: 'Confirm to delete',
+      message: 'Are you sure you want to delete?',
+      buttons: [
+        {
+          label: 'Yes',
+          onClick: () => {
+            deleteAdmin(admin.userId);
+            handleAlert();
+            setTimeout(() => {
+              window.location.reload();
+            }, 3000);
+            console.log('Deleted');
+          }
+        },
+        {
+          label: 'No',
+          onClick: () => {
+            // Cancel delete operation
+            console.log('Deletion cancelled');
+          }
+        }
+      ]
+    });
+  };
