+++ frontend/src/components/AdminPageComp/AdminMovie.jsx
@@ -2,6 +2,9 @@ import { useEffect, useState } from "react"
 import ViewPopup from "../ViewPopup";
 import axios from "axios";
 import SimpleAlert from "../SimpleAlert";
+import { confirmAlert } from 'react-confirm-alert';
+import 'react-confirm-alert/src/react-confirm-alert.css';
+
 
 const AdminMovie = ({movie}) => {
 
@@ -9,6 +12,7 @@ const AdminMovie = ({movie}) => {
   const [viewSynopsis, setViewSynopsis] = useState(false);
   const [showAlert, setShowAlert] = useState(false);
   const [showTimes, setShowTimes] = useState([]);
+  
 
   // Generate reviewText outside the if statement to make it accessible in JSX
   const reviewText = movie.reviews ? 
@@ -60,13 +64,37 @@ const AdminMovie = ({movie}) => {
     setTimeout(() => setShowAlert(false), 3000);
   };
 
+
+
   const handleDelete = () => {
-    deleteMovie(movie.id);
-    handleAlert();
-    setTimeout(() => {
-      window.location.reload();
-    }, 3000);
-  }
+    confirmAlert({
+      title: 'Confirm to delete',
+      message: 'Are you sure you want to delete?',
+      buttons: [
+        {
+          label: 'Yes',
+          onClick: () => {
+            deleteMovie(movie.id);
+            handleAlert();
+            setTimeout(() => {
+              window.location.reload();
+            }, 3000);
+            console.log('Deleted');
+          }
+        },
+        {
+          label: 'No',
+          onClick: () => {
+            // Cancel delete operation
+            console.log('Deletion cancelled');
+          }
+        }
+      ]
+    });
+  };
+  
+
+  
 
   return (
     <div className="admin__movie">
